Move placeholder validation and tests to HED 3 part of code
