+++ validator/event/hed3.js
@@ -4,8 +4,8 @@ import { IssueError } from '../../common/issues/issues'
 import ParsedHedGroup from '../parser/parsedHedGroup'
 import { ParsedHedTag } from '../parser/parsedHedTag'
 import { getParsedParentTags } from '../../utils/hedData'
-import { getParentTag, getTagName, hedStringIsAGroup } from '../../utils/hedStrings'
-import { isNumber } from '../../utils/string'
+import { getParentTag, getTagName, hedStringIsAGroup, replaceTagNameWithPound } from '../../utils/hedStrings'
+import { getCharacterCount, isNumber } from '../../utils/string'
 import { HedValidator } from './validator'
 
 const tagGroupType = 'tagGroup'
@@ -30,9 +30,21 @@ export class Hed3Validator extends HedValidator {
     this.definitions = definitions
   }
 
+  validateStringLevel() {
+    super.validateStringLevel()
+    this.validateFullParsedHedString()
+  }
+
   validateEventLevel() {
     super.validateEventLevel()
     this.validateTopLevelTagGroups()
+  }
+
+  /**
+   * Validate the full parsed HED string.
+   */
+  validateFullParsedHedString() {
+    this.checkPlaceholderStringSyntax()
   }
 
   /**
@@ -43,6 +55,9 @@ export class Hed3Validator extends HedValidator {
     if (this.definitions !== null) {
       this.checkForMissingDefinitions(tag, 'Def')
       this.checkForMissingDefinitions(tag, 'Def-expand')
+    }
+    if (this.options.expectValuePlaceholderString) {
+      this.checkPlaceholderTagSyntax(tag)
     }
   }
 
@@ -81,6 +96,29 @@ export class Hed3Validator extends HedValidator {
   }
 
   /**
+   * Check if an individual HED tag is in the schema or is an allowed extension.
+   */
+  checkIfTagIsValid(tag, previousTag) {
+    if (tag.existsInSchema || tag.takesValue) {
+      return
+    }
+
+    if (this.options.expectValuePlaceholderString && getCharacterCount(tag.formattedTag, '#') === 1) {
+      const valueTag = replaceTagNameWithPound(tag.formattedTag)
+      if (getCharacterCount(valueTag, '#') === 1) {
+        // Ending placeholder was replaced with itself.
+        this.pushIssue('invalidPlaceholder', {
+          tag: tag,
+        })
+      } /* else {
+        Handled in checkPlaceholderTagSyntax().
+      } */
+    } else {
+      super.checkIfTagIsValid(tag, previousTag)
+    }
+  }
+
+  /**
    * Check that the unit is valid for the tag's unit class.
    * @param {ParsedHed3Tag} tag A HED tag.
    */
@@ -106,6 +144,115 @@ export class Hed3Validator extends HedValidator {
       if (!validValue) {
         this.pushIssue('invalidValue', { tag: tag })
       }
+    }
+  }
+
+  /**
+   * Check basic placeholder tag syntax.
+   *
+   * @param {ParsedHedTag} tag A HED tag.
+   */
+  checkPlaceholderTagSyntax(tag) {
+    const placeholderCount = getCharacterCount(tag.formattedTag, '#')
+    if (placeholderCount === 1) {
+      const valueTag = replaceTagNameWithPound(tag.formattedTag)
+      if (getCharacterCount(valueTag, '#') !== 1) {
+        this.pushIssue('invalidPlaceholder', {
+          tag: tag,
+        })
+      }
+    } else if (placeholderCount > 1) {
+      // More than one placeholder.
+      this.pushIssue('invalidPlaceholder', {
+        tag: tag,
+      })
+    }
+  }
+
+  /**
+   * Check full-string placeholder syntax.
+   */
+  checkPlaceholderStringSyntax() {
+    const standalonePlaceholders = {
+      // Count of placeholders not in Definition groups.
+      placeholders: 0,
+      // Whether an Issue has already been generated for an excess placeholder outside a Definition group.
+      issueGenerated: false,
+    }
+    this._checkStandalonePlaceholderStringSyntaxInGroup(this.parsedString.topLevelTags, standalonePlaceholders)
+    // Loop over the top-level tag groups.
+    for (const tagGroup of this.parsedString.tagGroups) {
+      if (tagGroup.isDefinitionGroup) {
+        this._checkDefinitionPlaceholderStringSyntaxInGroup(tagGroup)
+      } else if (!standalonePlaceholders.issueGenerated) {
+        this._checkStandalonePlaceholderStringSyntaxInGroup(tagGroup.tagIterator(), standalonePlaceholders)
+      }
+    }
+    if (this.options.expectValuePlaceholderString && standalonePlaceholders.placeholders === 0) {
+      this.pushIssue('missingPlaceholder', {
+        string: this.parsedString.hedString,
+      })
+    }
+  }
+
+  /**
+   * Check Definition-related placeholder syntax in a tag group.
+   *
+   * @param {ParsedHedGroup} tagGroup A HED tag group.
+   * @private
+   */
+  _checkDefinitionPlaceholderStringSyntaxInGroup(tagGroup) {
+    // Count of placeholders within this Definition group.
+    let definitionPlaceholders = 0
+    const definitionHasPlaceholder = tagGroup.definitionValue === '#'
+    const definitionName = tagGroup.definitionName
+    for (const tag of tagGroup.tagIterator()) {
+      if (!definitionHasPlaceholder || tag !== tagGroup.definitionTag) {
+        definitionPlaceholders += getCharacterCount(tag.formattedTag, '#')
+      }
+    }
+    const isValid =
+      (!definitionHasPlaceholder && definitionPlaceholders === 0) ||
+      (definitionHasPlaceholder && definitionPlaceholders === 1)
+    if (!isValid) {
+      this.pushIssue('invalidPlaceholderInDefinition', {
+        definition: definitionName,
+      })
+    }
+  }
+
+  /**
+   * Check non-Definition-related placeholder syntax in a tag group.
+   *
+   * @param {ParsedHedTag[]|Generator<ParsedHedTag>} tags A HED tag iterator.
+   * @param {{placeholders: number, issueGenerated: boolean}} standalonePlaceholders The validator's standalone placeholder context.
+   * @private
+   */
+  _checkStandalonePlaceholderStringSyntaxInGroup(tags, standalonePlaceholders) {
+    let firstStandaloneTag
+    for (const tag of tags) {
+      const tagString = tag.formattedTag
+      const tagPlaceholders = getCharacterCount(tagString, '#')
+      standalonePlaceholders.placeholders += tagPlaceholders
+      if (!firstStandaloneTag && tagPlaceholders > 0) {
+        firstStandaloneTag = tag
+      }
+      if (
+        tagPlaceholders === 0 ||
+        (standalonePlaceholders.placeholders <= 1 &&
+          (this.options.expectValuePlaceholderString || standalonePlaceholders.placeholders === 0))
+      ) {
+        continue
+      }
+      if (this.options.expectValuePlaceholderString && !standalonePlaceholders.issueGenerated) {
+        this.pushIssue('invalidPlaceholder', {
+          tag: firstStandaloneTag,
+        })
+      }
+      this.pushIssue('invalidPlaceholder', {
+        tag: tag,
+      })
+      standalonePlaceholders.issueGenerated = true
     }
   }
